Add Yarn Output Production

+++ Com.Danliris.Service.Spinning.Lib/Models/Yarn.cs
@@ -11,7 +11,7 @@ namespace Com.Danliris.Service.Spinning.Lib.Models
     {
         public string Code { get; set; }
         public string Name { get; set; }
-        public decimal Ne { get; set; }
+        public double Ne { get; set; }
         public string Remark { get; set; }
 
         public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
